+++ src/SFA.DAS.EAS.Web.IntegrationTests/TestUtils/DataAccess/EmployerAccountsDbBuilder.cs
@@ -4,8 +4,8 @@ using SFA.DAS.EAS.Account.API.IntegrationTests.ModelBuilders;
 using SFA.DAS.EAS.Account.API.IntegrationTests.TestUtils.DataAccess.Adapters;
 using SFA.DAS.EAS.Account.API.IntegrationTests.TestUtils.DataAccess.Dtos;
 using SFA.DAS.EAS.Infrastructure.Data;
-using SFA.DAS.Hashing;
 using SFA.DAS.HashingService;
+using SFA.DAS.ObsoleteHashing;
 
 namespace SFA.DAS.EAS.Account.API.IntegrationTests.TestUtils.DataAccess
 {
