+++ src/SFA.DAS.EmployerAccounts.Web.UnitTests/Orchestrators/OrganisationOrchestratorTests/WhenICreateALegalEntityOfOtherType.cs
@@ -4,12 +4,11 @@ using MediatR;
 using Moq;
 using NUnit.Framework;
 using SFA.DAS.EmployerAccounts.Interfaces;
+using SFA.DAS.EmployerAccounts.MarkerInterfaces;
 using SFA.DAS.EmployerAccounts.Models.Account;
 using SFA.DAS.EmployerAccounts.Web.Orchestrators;
 using SFA.DAS.EmployerAccounts.Web.ViewModels;
-using SFA.DAS.Hashing;
 using SFA.DAS.NLog.Logger;
-
 namespace SFA.DAS.EmployerAccounts.Web.UnitTests.Orchestrators.OrganisationOrchestratorTests
 {
     public class WhenICreateALegalEntityOfOtherType
