+++ src/SFA.DAS.EmployerAccounts.Web.UnitTests/Orchestrators/OrganisationOrchestratorTests/WhenISearchForALegalEntity.cs
@@ -4,10 +4,10 @@ using MediatR;
 using Moq;
 using NUnit.Framework;
 using SFA.DAS.EmployerAccounts.Interfaces;
+using SFA.DAS.EmployerAccounts.MarkerInterfaces;
 using SFA.DAS.EmployerAccounts.Models.Account;
 using SFA.DAS.EmployerAccounts.Queries.GetAccountLegalEntities;
 using SFA.DAS.EmployerAccounts.Web.Orchestrators;
-using SFA.DAS.Hashing;
 using SFA.DAS.NLog.Logger;
 
 namespace SFA.DAS.EmployerAccounts.Web.UnitTests.Orchestrators.OrganisationOrchestratorTests
