+++ src/SFA.DAS.EmployerAccounts/MarkerInterfaces/IAccountLegalEntityPublicHashingService.cs
@@ -1,0 +1,9 @@
+﻿using SFA.DAS.HashingService;
+
+namespace SFA.DAS.EmployerAccounts.MarkerInterfaces
+{
+    public interface IAccountLegalEntityPublicHashingService : IHashingService
+    {
+        
+    }
+}
