+++ src/SFA.DAS.EmployerApprenticeshipsService.Application.UnitTests/Queries/GetTransferTransactionDetailsTests/WhenIGetsTransferTransactionDetails.cs
@@ -9,7 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using SFA.DAS.Hashing;
+using SFA.DAS.ObsoleteHashing;
 
 namespace SFA.DAS.EAS.Application.UnitTests.Queries.GetTransferTransactionDetailsTests
 {
