+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Commands/SendTransferConnectionInvitation/SendTransferConnectionInvitationCommandHandler.cs
@@ -2,7 +2,7 @@ using System.Threading.Tasks;
 using MediatR;
 using SFA.DAS.EAS.Domain.Data.Repositories;
 using SFA.DAS.EAS.Domain.Interfaces;
-using SFA.DAS.Hashing;
+using SFA.DAS.ObsoleteHashing;
 
 namespace SFA.DAS.EAS.Application.Commands.SendTransferConnectionInvitation
 {
