+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Queries/GetEmployerAccountTransactions/GetEmployerAccountTransactionsHandler.cs
@@ -12,7 +12,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using SFA.DAS.EAS.Domain.Models.ExpiredFunds;
-using SFA.DAS.Hashing;
+using SFA.DAS.ObsoleteHashing;
 
 namespace SFA.DAS.EAS.Application.Queries.GetEmployerAccountTransactions
 {
