+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Queries/GetTransferTransactionDetails/GetTransferTransactionDetailsQueryHandler.cs
@@ -6,7 +6,7 @@ using SFA.DAS.EAS.Infrastructure.Extensions;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using SFA.DAS.Hashing;
+using SFA.DAS.ObsoleteHashing;
 
 namespace SFA.DAS.EAS.Application.Queries.GetTransferTransactionDetails
 {
