+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Queries/SendTransferConnectionInvitation/SendTransferConnectionInvitationQueryHandler.cs
@@ -7,7 +7,7 @@ using MediatR;
 using SFA.DAS.EAS.Application.Dtos;
 using SFA.DAS.EAS.Domain.Models.TransferConnections;
 using SFA.DAS.EAS.Infrastructure.Data;
-using SFA.DAS.Hashing;
+using SFA.DAS.ObsoleteHashing;
 using SFA.DAS.Validation;
 
 namespace SFA.DAS.EAS.Application.Queries.SendTransferConnectionInvitation
