+++ src/SFA.DAS.EmployerApprenticeshipsService.Infrastructure/Data/AccountRepository.cs
@@ -12,7 +12,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Linq;
 using System.Threading.Tasks;
-using SFA.DAS.Hashing;
+using SFA.DAS.ObsoleteHashing;
 
 namespace SFA.DAS.EAS.Infrastructure.Data
 {
