+++ src/SFA.DAS.EmployerFinance.UnitTests/Queries/GetTransferTransactionDetailsTests/WhenIGetsTransferTransactionDetails.cs
@@ -5,12 +5,12 @@ using SFA.DAS.EmployerFinance.Models.Payments;
 using SFA.DAS.EmployerFinance.Models.Transaction;
 using SFA.DAS.EmployerFinance.Models.Transfers;
 using SFA.DAS.EmployerFinance.Queries.GetTransferTransactionDetails;
-using SFA.DAS.Hashing;
 using SFA.DAS.Testing.EntityFramework;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using SFA.DAS.ObsoleteHashing;
 
 namespace SFA.DAS.EmployerFinance.UnitTests.Queries.GetTransferTransactionDetailsTests
 {
