+++ src/SFA.DAS.EmployerFinance/Queries/GetTransferTransactionDetails/GetTransferTransactionDetailsQueryHandler.cs
@@ -3,10 +3,10 @@ using SFA.DAS.EmployerFinance.Data;
 using SFA.DAS.EmployerFinance.Extensions;
 using SFA.DAS.EmployerFinance.Models.Transaction;
 using SFA.DAS.EmployerFinance.Models.Transfers;
-using SFA.DAS.Hashing;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using SFA.DAS.ObsoleteHashing;
 
 namespace SFA.DAS.EmployerFinance.Queries.GetTransferTransactionDetails
 {
