+++ src/SFA.DAS/ObsoleteHashing/IAccountLegalEntityPublicHashingService.cs
@@ -1,6 +1,6 @@
 ﻿using SFA.DAS.HashingService;
 
-namespace SFA.DAS.Hashing
+namespace SFA.DAS.ObsoleteHashing
 {
     public interface IAccountLegalEntityPublicHashingService : IHashingService
     {
