+++ src/SFA.DAS/ObsoleteHashing/IPublicHashingService.cs
@@ -1,6 +1,6 @@
 ﻿using SFA.DAS.HashingService;
 
-namespace SFA.DAS.Hashing
+namespace SFA.DAS.ObsoleteHashing
 {
     public interface IPublicHashingService : IHashingService
     {
